requirements.txt
asgiref==3.5.2
autopep8==1.6.0
blinker==1.5
certifi==2022.6.15
charset-normalizer==2.1.1
crispy-bootstrap5==0.6
Django==4.0.4
django-browser-reload==1.6.0
django-crispy-forms==1.14.0
django-filter==22.1
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
django-user-sessions==1.7.1
django-widget-tweaks==1.4.12
dnspython==2.2.1
falcon==3.1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3Sss
MarkupSafe==2.1.1
mongoengine==0.24.2
mssql-django==1.1.3
mysqlclient==2.1.0
Pillow==9.1.1
pycodestyle==2.8.0
pymongo==4.2.0
pyodbc==4.0.32
pytz==2022.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.12
Werkzeug==2.2.2
windows-curses==2.3.0
